thumbnails have fixed height, tool to generate only missing thumbnails

+++ biography-core/src/main/java/de/zaunkoenigweg/biography/core/archive/ArchiveImportService.java
@@ -141,7 +141,7 @@ public class ArchiveImportService {
         metadataService.writeMetadataToJsonFile(jsonFile, metadata);
     }
     
-    public boolean generateThumbnails(File file) {
+    public boolean generateThumbnails(File file, boolean force) {
     	
     	File thumbsFolder200 = new File(this.thumbnailsFolder, "200");
     	File thumbsFolder300 = new File(this.thumbnailsFolder, "300");
@@ -151,12 +151,19 @@ public class ArchiveImportService {
 			client = new DefaultHttpClient();
 
 			// TODO method for one resizing w/param (200, 300, ...)
-			String thumborUri = this.thumborUrl + "fit-in/200x200/filters:fill(black)/" + archiveFolder.toPath().relativize(file.toPath()).toString();
+
+			File file200 = BiographyFileUtils.getArchiveFileFromShortFilename(thumbsFolder200, file.getName());
+			File file300 = BiographyFileUtils.getArchiveFileFromShortFilename(thumbsFolder300, file.getName());
+
+			if(!force && (file200.exists() && file300.exists())) {
+				return true;
+			}
+			
+			String thumborUri = this.thumborUrl + "0x200/" + archiveFolder.toPath().relativize(file.toPath()).toString();
 			HttpGet request = new HttpGet(thumborUri);
 			HttpResponse response = client.execute(request);
 			if(response.getStatusLine().getStatusCode()==200) {
 				BufferedInputStream bis = new BufferedInputStream(response.getEntity().getContent());
-				File file200 = BiographyFileUtils.getArchiveFileFromShortFilename(thumbsFolder200, file.getName());
 				file200.getParentFile().mkdirs();
 				BufferedOutputStream bos = new BufferedOutputStream(new FileOutputStream(file200));
 				int inByte;
@@ -164,12 +171,12 @@ public class ArchiveImportService {
 				bis.close();
 				bos.close();
 			}
-			thumborUri = this.thumborUrl + "fit-in/300x300/filters:fill(black)/" + archiveFolder.toPath().relativize(file.toPath()).toString();
+
+			thumborUri = this.thumborUrl + "0x300/" + archiveFolder.toPath().relativize(file.toPath()).toString();
 			request = new HttpGet(thumborUri);
 			response = client.execute(request);
 			if(response.getStatusLine().getStatusCode()==200) {
 				BufferedInputStream bis = new BufferedInputStream(response.getEntity().getContent());
-				File file300 = BiographyFileUtils.getArchiveFileFromShortFilename(thumbsFolder300, file.getName());
 				file300.getParentFile().mkdirs();
 				BufferedOutputStream bos = new BufferedOutputStream(new FileOutputStream(file300));
 				int inByte;
